rename filename of the content exclusion custom hook
